wishlist-ai-backend/pyproject.toml
[project]
name = "wishlist-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.133.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
